Fix memory issue when distributing work
apply_async is not lazy, therefore there are still memory issues when
distributing work to the pool. I now use imap. Its not ideal because
there is currently no way to monitor the progress queues and the status
of the pool. Future work should look into the writing a layer of the
mp.Pool with apply async that lazy loads data from the generator, and
uses the Async Results to determine when the total work is done.